Rename groupnorm to groupnorm fwd in example
