+++ example/42_groupnorm_fwd/common.hpp
